+++ man-roxygen/arg_iraceresults.R
@@ -1,0 +1,2 @@
+#' @param iraceResults (`list()`|`character(1)`) `iraceResults` object created by \pkg{irace} and typically saved in the log file `irace.Rdata`. Alternatively, the path to the log file from which the `iraceResults` object will be loaded.
+#' @md